Move FluentAPI test cases to separate file
