+++ analyzers/tests/SonarAnalyzer.UnitTest/Rules/TimestampsShouldNotBeUsedAsPrimaryKeysTest.cs
@@ -18,7 +18,6 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-using System.IO;
 using SonarAnalyzer.Rules.CSharp;
 
 namespace SonarAnalyzer.UnitTest.Rules;
@@ -26,48 +25,43 @@ namespace SonarAnalyzer.UnitTest.Rules;
 [TestClass]
 public class TimestampsShouldNotBeUsedAsPrimaryKeysTest
 {
+    private readonly VerifierBuilder verifierCS = CreateVerifier<TimestampsShouldNotBeUsedAsPrimaryKeys>();
+
     [TestMethod]
     public void TimestampsShouldNotBeUsedAsPrimaryKeys_CSharp() =>
-        CreateVerifier<TimestampsShouldNotBeUsedAsPrimaryKeys>("TimestampsShouldNotBeUsedAsPrimaryKeys.CSharp.cs").Verify();
+        verifierCS.AddPaths("TimestampsShouldNotBeUsedAsPrimaryKeys.CSharp.cs").Verify();
 
     [TestMethod]
     public void TimestampsShouldNotBeUsedAsPrimaryKeys_CSharp9() =>
-        CreateVerifier<TimestampsShouldNotBeUsedAsPrimaryKeys>("TimestampsShouldNotBeUsedAsPrimaryKeys.CSharp9.cs")
+        verifierCS
+            .AddPaths("TimestampsShouldNotBeUsedAsPrimaryKeys.CSharp9.cs")
             .WithOptions(ParseOptionsHelper.FromCSharp9)
             .Verify();
 
     [TestMethod]
     public void TimestampsShouldNotBeUsedAsPrimaryKeys_NoReferenceToEntityFramework_CSharp() =>
-        new VerifierBuilder<TimestampsShouldNotBeUsedAsPrimaryKeys>()
-            .AddPaths("TimestampsShouldNotBeUsedAsPrimaryKeys.NoReferenceToEntityFramework.CSharp.cs")
-            .Verify();
+        new VerifierBuilder<TimestampsShouldNotBeUsedAsPrimaryKeys>().AddPaths("TimestampsShouldNotBeUsedAsPrimaryKeys.NoReferenceToEntityFramework.CSharp.cs").Verify();
 
 #if NET
     [TestMethod]
     public void TimestampsShouldNotBeUsedAsPrimaryKeys_EntityFrameworkCore_CSharp() =>
-        CreateVerifier<TimestampsShouldNotBeUsedAsPrimaryKeys>("TimestampsShouldNotBeUsedAsPrimaryKeys.EntityFrameworkCore.CSharp.cs").Verify();
+        verifierCS.AddPaths("TimestampsShouldNotBeUsedAsPrimaryKeys.EntityFrameworkCore.CSharp.cs").Verify();
+
+    [TestMethod]
+    public void TimestampsShouldNotBeUsedAsPrimaryKeys_FluentApi_CSharp() =>
+        verifierCS.AddPaths("TimestampsShouldNotBeUsedAsPrimaryKeys.FluentApi.CSharp.cs").Verify();
 #endif
 
-    private static VerifierBuilder CreateVerifier<TAnalyzer>(string testFileName)
-        where TAnalyzer : DiagnosticAnalyzer, new()
-    {
-        var builder = new VerifierBuilder<TAnalyzer>()
+    private static VerifierBuilder CreateVerifier<TAnalyzer>()
+        where TAnalyzer : DiagnosticAnalyzer, new() =>
+        new VerifierBuilder<TAnalyzer>()
             .AddReferences(NuGetMetadataReference.SystemComponentModelAnnotations())
 
 #if NET
             .AddReferences(NuGetMetadataReference.MicrosoftEntityFrameworkCore("7.0.0"))
             .AddReferences(NuGetMetadataReference.MicrosoftEntityFrameworkCoreAbstractions("7.0.0"));
-        return AddTestFileWithExtraLinePrepended(builder, testFileName, "using Microsoft.EntityFrameworkCore;");
 #else
             .AddReferences(NuGetMetadataReference.MicrosoftEntityFramework("6.0.0"));
-        return AddTestFileWithExtraLinePrepended(builder, testFileName, "using System.Data.Entity;using ModelBuilder = System.Data.Entity.DbModelBuilder;");
 #endif
 
-    }
-
-    private static VerifierBuilder AddTestFileWithExtraLinePrepended(VerifierBuilder builder, string testFileName, string extraLine)
-    {
-        var content = File.ReadAllText($@"TestCases\{testFileName}");
-        return builder.AddSnippet($"{extraLine}{Environment.NewLine}{content}", testFileName);
-    }
 }
