+++ analyzers/tests/SonarAnalyzer.UnitTest/TestCases/TimestampsShouldNotBeUsedAsPrimaryKeys.CSharp.cs
@@ -174,43 +174,3 @@ class NonClassTypes
         DateTime Id { get; set; }                       // Compliant - issue will be raised in the implementing class
     }
 }
-
-class FluentApi
-{
-    class PersonDbContext: DbContext
-    {
-        public DbSet<Person> People { get; set; }
-
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            modelBuilder.Entity<Person>()
-                .HasKey(x => x.DateOfBirth);
-        }
-    }
-
-    class Person
-    {
-        public DateTime DateOfBirth { get; set; }       // FN - keys created with the Fluent API are too complex to track
-    }
-}
-
-class BuildErrors
-{
-    class AttributeDoesNotExist
-    {
-        []                                              // Error [CS1001]
-        public DateTime Identifier { get; set; }
-    }
-
-    class MissingPropertyType
-    {
-        [Key]
-        public Identifier { get; set; }                 // Error [CS1519, CS1519, CS1519, CS1519, CS1519, CS1519]
-    }
-
-    class MissingPropertyName
-    {
-        [Key]
-        public DateTime { get; set; }                   // Error [CS1519, CS1519, CS1519, CS1519, CS1519, CS1519]
-    }                                                   // Error [CS1022]
-}                                                       // Error [CS1022]
